fix(extensions): complete turn tool migration
