+++ codex-rs/core/Cargo.toml
@@ -136,6 +136,7 @@ assert_matches = { workspace = true }
 codex-image-generation-extension = { workspace = true }
 codex-home = { workspace = true }
 codex-otel = { workspace = true }
+codex-skill-search-extension = { workspace = true }
 codex-test-binary-support = { workspace = true }
 codex-utils-cargo-bin = { workspace = true }
 codex-web-search-extension = { workspace = true }
