+++ codex-rs/core/src/session/review.rs
@@ -105,6 +105,7 @@ pub(super) async fn spawn_review_thread(
     extension_data.insert(HostLoadedSkills::new(
         parent_turn_context.turn_skills.outcome.clone(),
     ));
+
     let review_turn_context = TurnContext {
         sub_id: review_turn_id.clone(),
         trace_id: current_span_trace_id(),
