+++ codex-rs/ext/skill-search/src/extension.rs
@@ -5,6 +5,7 @@ use codex_core_skills::HostLoadedSkills;
 use codex_extension_api::ConfigContributor;
 use codex_extension_api::ContextContributor;
 use codex_extension_api::ExtensionData;
+use codex_extension_api::ExtensionFuture;
 use codex_extension_api::ExtensionRegistryBuilder;
 use codex_extension_api::PromptFragment;
 use codex_extension_api::ThreadLifecycleContributor;
@@ -39,7 +40,7 @@ impl ContextContributor for SkillSearchExtension {
         &'a self,
         _session_store: &'a ExtensionData,
         thread_store: &'a ExtensionData,
-    ) -> std::pin::Pin<Box<dyn std::future::Future<Output = Vec<PromptFragment>> + Send + 'a>> {
+    ) -> ExtensionFuture<'a, Vec<PromptFragment>> {
         Box::pin(async move {
             let Some(config) = thread_store.get::<SkillSearchExtensionConfig>() else {
                 return Vec::new();
@@ -112,56 +113,5 @@ pub fn install(registry: &mut ExtensionRegistryBuilder<Config>) {
 }
 
 #[cfg(test)]
-mod tests {
-    use codex_extension_api::PromptSlot;
-    use codex_extension_api::ToolContributor;
-    use codex_extension_api::ToolName;
-    use pretty_assertions::assert_eq;
-
-    use super::*;
-    use crate::tool::SKILL_SEARCH_TOOL_NAME;
-
-    #[tokio::test]
-    async fn prompt_contribution_is_gated_by_feature_config() {
-        let extension = SkillSearchExtension;
-        let session_store = ExtensionData::new("session");
-        let thread_store = ExtensionData::new("thread");
-
-        assert!(
-            extension
-                .contribute(&session_store, &thread_store)
-                .await
-                .is_empty()
-        );
-
-        thread_store.insert(SkillSearchExtensionConfig { enabled: true });
-        let fragments = extension.contribute(&session_store, &thread_store).await;
-
-        assert_eq!(fragments.len(), 1);
-        assert_eq!(fragments[0].slot(), PromptSlot::DeveloperPolicy);
-        assert!(fragments[0].text().contains(SKILL_SEARCH_TOOL_NAME));
-    }
-
-    #[test]
-    fn tool_contribution_is_gated_by_feature_config() {
-        let extension = SkillSearchExtension;
-        let session_store = ExtensionData::new("session");
-        let thread_store = ExtensionData::new("thread");
-        let turn_store = ExtensionData::new("turn");
-
-        assert!(
-            extension
-                .tools(&session_store, &thread_store, &turn_store)
-                .is_empty()
-        );
-
-        thread_store.insert(SkillSearchExtensionConfig { enabled: true });
-        let tool_names = extension
-            .tools(&session_store, &thread_store, &turn_store)
-            .into_iter()
-            .map(|tool| tool.tool_name())
-            .collect::<Vec<_>>();
-
-        assert_eq!(tool_names, vec![ToolName::plain(SKILL_SEARCH_TOOL_NAME)]);
-    }
-}
+#[path = "extension_tests.rs"]
+mod tests;
